Rename main module file to meet PEP-8 conventions.
